+++ assets/prototype/application/valueObjects/config/index.js
@@ -3,6 +3,7 @@
  */
 import { Currency, CurrencyProps } from './Currency';
 import { CurrentUser, CurrentUserProps } from './CurrentUser';
+import { DateTimeFormats, DateTimeFormatsProps } from './DateTimeFormats';
 import { Locale, LocaleProps } from './Locale';
 import { SiteUrl, SiteUrlProps } from './SiteUrl';
 import { Timezone, TimezoneProps } from './Timezone';
@@ -12,6 +13,8 @@ export {
 	CurrencyProps,
 	CurrentUser,
 	CurrentUserProps,
+	DateTimeFormats,
+	DateTimeFormatsProps,
 	Locale,
 	LocaleProps,
 	SiteUrl,
